アプリ側でマッチさせる

+++ home/isucon/webapp/node/src/handlers/livecomment-handler.ts
@@ -332,37 +332,32 @@ export const moderateHandler = [
         )
         .catch(throwErrorWith('failed to get NG words'))
 
-      // NGワードにヒットする過去の投稿も全削除する
-      for (const ngword of ngwords) {
-        // ライブコメント一覧取得
-        const [livecomments] = await conn
+      // ライブコメント一覧取得
+      const [livecomments] = await conn
           .query<(LivecommentsModel & RowDataPacket)[]>(
             'SELECT * FROM livecomments WHERE livestream_id = ?',
             [livestreamId],
           )
           .catch(throwErrorWith('failed to get livecomments'))
 
+      for (const ngword of ngwords) {
         for (const livecomment of livecomments) {
+          if (livecomment.comment.indexOf(ngword.word)!==-1){
           await conn
             .query(
               `
                 DELETE FROM livecomments
                 WHERE
-                id = ? AND
-                (SELECT COUNT(*)
-                FROM
-                (SELECT ? AS text) AS texts
-                INNER JOIN
-                (SELECT CONCAT('%', ?, '%')	AS pattern) AS patterns
-                ON texts.text LIKE patterns.pattern) >= 1;
+                id = ?
               `,
-              [livecomment.id, livecomment.comment, ngword.word],
+              [livecomment.id],
             )
             .catch(
               throwErrorWith(
                 'failed to delete old livecomments that hit spams',
               ),
             )
+          }
         }
       }
